* changed FiringInfo::timeSent from a float to a double
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@20157 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/WorldWeapons.cxx
@@ -43,7 +43,7 @@ static int fireWorldWepReal(FlagType* type, float lifetime, PlayerId player,
 			    float dir, int shotID, float dt)
 {
   FiringInfo firingInfo;
-  firingInfo.timeSent = (float)TimeKeeper::getCurrent().getSeconds();
+  firingInfo.timeSent = TimeKeeper::getCurrent().getSeconds();
   firingInfo.flagType = type;
   firingInfo.lifetime = lifetime;
   firingInfo.shotType = type->flagShot;
@@ -73,7 +73,7 @@ static int fireWorldGMReal(FlagType* type, PlayerId targetPlayerID, float
                            float tilt, float dir, int shotID, float dt)
 {
     FiringInfo firingInfo;
-    firingInfo.timeSent = (float)TimeKeeper::getCurrent().getSeconds();
+    firingInfo.timeSent = TimeKeeper::getCurrent().getSeconds();
     firingInfo.flagType = type;
     firingInfo.lifetime = lifetime;
     firingInfo.shot.player = player;
